# docs/api/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DocReader PoC API",
    "description": "Local proof of concept for document ingestion, OCR, classification and extraction.\n\n**There is no authentication.** Every endpoint is anonymous, the Swagger page is public and\nno bearer token is required. Do not expose this API to the internet and do not send real\npersonal documents to it. Use synthetic, masked or authorized files only.\n\nConventions: JSON in camelCase, timestamps in UTC ISO 8601, errors as\n`application/problem+json`, and an `X-Correlation-Id` header on every response.\n\nStage 3 of the execution plan delivers upload, listing, detail, content, deletion, OCR\n(PP-OCRv5 on CPU, see ADR 0002), the raw text and the structured result of seven document\ntypes: CPF card, CIN/RG, CNH, proof of address, CNPJ card, CCMEI and social contract.\nOther document types are read as raw text only.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/webhook-subscriptions": {
      "post": {
        "tags": [
          "WebhookSubscriptions"
        ],
        "summary": "Creates a webhook subscription.",
        "description": "The secret is stored encrypted and no endpoint returns it. Send your own (16 to 256 characters) or omit it: a random one is\ngenerated and returned in this answer, and only in this answer. The URL must be public unless the deployment allows private\nnetworks (`WEBHOOK_ALLOW_PRIVATE_NETWORKS`).",
        "requestBody": {
          "description": "URL, events, optional product filter and optional secret.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookSubscriptionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookSubscriptionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/webhook-subscriptions",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "201": {
            "description": "Created. The `Location` header points at it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSubscriptionCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "The URL, the events or the secret are invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/validation-failed",
                  "title": "One or more validation errors occurred",
                  "status": 400,
                  "detail": "Check the errors property for the fields that were refused.",
                  "instance": "/api/v1/webhook-subscriptions",
                  "errorCode": "VALIDATION_FAILED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "422": {
            "description": "The product or service does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unprocessable-content",
                  "title": "Content cannot be processed",
                  "status": 422,
                  "detail": "There is no product or service with the id 0199c1f0-1111-7a10-9c44-2f1d8e6b4a21.",
                  "instance": "/api/v1/webhook-subscriptions",
                  "errorCode": "PRODUCT_SERVICE_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "WebhookSubscriptions"
        ],
        "summary": "Lists the webhook subscriptions. The secret is never included.",
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "productServiceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 500,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/webhook-subscriptions",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "200": {
            "description": "One page of subscriptions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSubscriptionResponsePagedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhook-subscriptions/{id}": {
      "get": {
        "tags": [
          "WebhookSubscriptions"
        ],
        "summary": "Returns one webhook subscription. The secret is never included.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identity.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/webhook-subscriptions/{id}",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "200": {
            "description": "The subscription.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSubscriptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "There is none with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/document-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No webhook subscription matches the requested identifier.",
                  "instance": "/api/v1/webhook-subscriptions/{id}",
                  "errorCode": "WEBHOOK_SUBSCRIPTION_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "WebhookSubscriptions"
        ],
        "summary": "Replaces the URL, the events, the product filter and the active flag of a subscription.",
        "description": "Send a `secret` to rotate the signing secret; omit it to keep the current one.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identity.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The new values.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhookSubscriptionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhookSubscriptionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhookSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/webhook-subscriptions/{id}",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "200": {
            "description": "Updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSubscriptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "The URL, the events or the secret are invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/validation-failed",
                  "title": "One or more validation errors occurred",
                  "status": 400,
                  "detail": "Check the errors property for the fields that were refused.",
                  "instance": "/api/v1/webhook-subscriptions/{id}",
                  "errorCode": "VALIDATION_FAILED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "404": {
            "description": "There is none with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/document-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No webhook subscription matches the requested identifier.",
                  "instance": "/api/v1/webhook-subscriptions/{id}",
                  "errorCode": "WEBHOOK_SUBSCRIPTION_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "422": {
            "description": "The product or service does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unprocessable-content",
                  "title": "Content cannot be processed",
                  "status": 422,
                  "detail": "The document has 84 pages and the limit is 50 pages.",
                  "instance": "/api/v1/webhook-subscriptions/{id}",
                  "errorCode": "PAGE_LIMIT_EXCEEDED",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "WebhookSubscriptions"
        ],
        "summary": "DESTRUCTIVE AND IRREVERSIBLE — Deletes a webhook subscription and what was still queued for it.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identity.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/unexpected-error",
                  "title": "Unexpected error",
                  "status": 500,
                  "detail": "The request could not be completed. Check the API logs using the correlation id of this response.",
                  "instance": "/api/v1/webhook-subscriptions/{id}",
                  "errorCode": "UNEXPECTED_ERROR",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          },
          "204": {
            "description": "Deleted."
          },
          "404": {
            "description": "There is none with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://docreader.local/problems/document-not-found",
                  "title": "Resource not found",
                  "status": 404,
                  "detail": "No webhook subscription matches the requested identifier.",
                  "instance": "/api/v1/webhook-subscriptions/{id}",
                  "errorCode": "WEBHOOK_SUBSCRIPTION_NOT_FOUND",
                  "correlationId": "9f1c1d0f2a5b4c8e9d7a6b5c4d3e2f10"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWebhookSubscriptionRequest": {
        "required": [
          "events",
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "type": "string",
            "description": "Absolute http or https URL, with no credentials in it. Private, loopback and link-local addresses are refused unless the deployment allows them."
          },
          "secret": {
            "maxLength": 256,
            "type": "string",
            "description": "The HMAC-SHA256 key, 16 to 256 characters. Leave it out to have one generated; it is then returned once, in this answer.",
            "nullable": true
          },
          "events": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "One or more of `document.completed`, `document.failed`, `document.purged`."
          },
          "productServiceId": {
            "type": "string",
            "description": "Only notify documents of this product or service; omit for every document.",
            "format": "uuid",
            "nullable": true
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProductServiceReferenceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identity.",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "description": "Unique code."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          }
        },
        "additionalProperties": false,
        "description": "The product or service a document is linked to."
      },
...
      "UpdateWebhookSubscriptionRequest": {
        "required": [
          "events",
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "maxLength": 2048,
            "minLength": 1,
            "type": "string",
            "description": "Absolute http or https URL, with no credentials in it."
          },
          "secret": {
            "maxLength": 256,
            "type": "string",
            "description": "A new signing secret, 16 to 256 characters. Omit it to keep the current one; there is no way to read the current one.",
            "nullable": true
          },
          "events": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "One or more of `document.completed`, `document.failed`, `document.purged`. Replaces the current list."
          },
          "productServiceId": {
            "type": "string",
            "description": "Only notify documents of this product or service; omit (null) for every document.",
            "format": "uuid",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "WebhookSubscriptionCreatedResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identity.",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "description": "Where the notifications are POSTed."
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Events it receives."
          },
          "productService": {
            "$ref": "#/components/schemas/ProductServiceReferenceResponse"
          },
          "active": {
            "type": "boolean",
            "description": "Whether it receives notifications."
          },
          "createdAt": {
            "type": "string",
            "description": "Creation instant, in UTC.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last change, in UTC.",
            "format": "date-time"
          },
          "secret": {
            "type": "string",
            "description": "The signing secret, only when none was sent in the request and one had to be generated. Save it now: it is stored encrypted and no endpoint returns it again. Null when you supplied your own.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The answer to creating a subscription: the subscription and, when the secret was generated, the secret itself, this once.",
        "example": {
          "id": "0199c1f0-4444-7a10-9c44-2f1d8e6b4a21",
          "url": "https://webhook.site/6c0e2f6a-1b7d-4c8e-9a5f-3d2b1c0e9f8a",
          "events": [
            "document.completed",
            "document.failed"
          ],
          "productService": {
            "id": "0199c1f0-1111-7a10-9c44-2f1d8e6b4a21",
            "code": "CONTA-PJ",
            "name": "Abertura de conta PJ"
          },
          "active": true,
          "createdAt": "2026-09-24T22:00:00Z",
          "updatedAt": "2026-09-24T22:00:00Z",
          "secret": "9f2c1ab5d0e74c3c8a6b21f04d5e9a77c3b1e8d24f6a90b5c7d1e3f8a2b4c6d0"
        }
      },
      "WebhookSubscriptionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identity.",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "description": "Where the notifications are POSTed."
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Events it receives: `document.completed`, `document.failed`, `document.purged`."
          },
          "productService": {
            "$ref": "#/components/schemas/ProductServiceReferenceResponse"
          },
          "active": {
            "type": "boolean",
            "description": "An inactive subscription receives nothing, and what was still queued for it is dropped."
          },
          "createdAt": {
            "type": "string",
            "description": "Creation instant, in UTC.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last change, in UTC.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "A subscriber that is told when a document reaches a final state. The signing secret is never returned.",
        "example": {
          "id": "0199c1f0-4444-7a10-9c44-2f1d8e6b4a21",
          "url": "https://webhook.site/6c0e2f6a-1b7d-4c8e-9a5f-3d2b1c0e9f8a",
          "events": [
            "document.completed",
            "document.failed"
          ],
          "productService": {
            "id": "0199c1f0-1111-7a10-9c44-2f1d8e6b4a21",
            "code": "CONTA-PJ",
            "name": "Abertura de conta PJ"
          },
          "active": true,
          "createdAt": "2026-09-24T22:00:00Z",
          "updatedAt": "2026-09-24T22:00:00Z"
        }
      },
      "WebhookSubscriptionResponsePagedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookSubscriptionResponse"
            },
            "description": "Items of the current page, newest first."
          },
          "page": {
            "type": "integer",
            "description": "One based page number.",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Number of items per page.",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "description": "Total number of items matching the filters.",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "description": "Number of pages for the current page size.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Envelope of every paginated listing.",
        "example": {
          "items": [
            {
              "id": "0199c1f0-4444-7a10-9c44-2f1d8e6b4a21",
              "url": "https://webhook.site/6c0e2f6a-1b7d-4c8e-9a5f-3d2b1c0e9f8a",
              "events": [
                "document.completed",
                "document.failed"
              ],
              "productService": {
                "id": "0199c1f0-1111-7a10-9c44-2f1d8e6b4a21",
                "code": "CONTA-PJ",
                "name": "Abertura de conta PJ"
              },
              "active": true,
              "createdAt": "2026-09-24T22:00:00Z",
              "updatedAt": "2026-09-24T22:00:00Z"
            }
          ],
          "page": 1,
          "pageSize": 20,
          "totalCount": 1,
          "totalPages": 1
        }
      }
    }
  },
...
}